Created function countTaxObligation, written test by using test driving development

+++ src/main/java/com/tvo/propertyregister/model/TaxRate.java
@@ -5,12 +5,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class TaxRate {
-
     private int id;
     private PropertyType propertyType;
-    private float tax;
+    private BigDecimal tax;
 }
